refactor(docx-reader): extract relationship path helpers into rels/paths.rs
